requirements.txt
babel==2.17.0
bcrypt==4.3.0
black==23.11.0
boto3==1.37.6
email_validator==2.2.0
fastapi==0.115.11
flake8==6.1.0
grpcio-status==1.71.0
h2==4.2.0
isort==5.12.0
langchain-community==0.3.27
langchain-google-genai==2.1.9
langchain-openai
langgraph-supervisor==0.0.29
passlib==1.7.4
PyJWT==2.10.1
pytest-asyncio==0.21.1
python-jose==3.4.0
python-json-logger==3.2.1
python-multipart==0.0.6
sentry-sdk==2.22.0
slack_bolt==1.23.0
sse-starlette==2.2.1
types-passlib==1.7.7.20241221
uvicorn==0.34.0
wheel==0.45.1
xxhash==3.5.0
python-dotenv==1.1.0
pydantic-settings==2.9.1
openai
APScheduler==3.11.0
beautifulsoup4==4.14.2
qdrant-client==1.15.1
jira==3.10.5
RapidFuzz==3.14.3